src/requirements.txt
Jinja2==3.1.6
pytest==8.3.5
bcrypt==4.3.0
pyzbar==0.1.9
opencv-python-headless==4.12.0.88
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
qrcode[pil]==8.2
mysql-connector-python==9.3.0
Flask-Migrate==4.1.0
